new(sinsp): support cgroups v2
Using a hardcoded subsystem list for cgroups v2 was a vast
simplification. For one thing, not all directories in the unified
hierarchy support all controllers, so it's totally possible
for a cgroup to only support e.g. `pids` and `memory`. In that case,
the settings for other controllers are taken from the parent cgroups.

So once we have the cgroup v2 path, we still need to walk up the cgroup
tree and check which subsystems are enabled for each particular tree
level, until we get all controllers or we reach the root of the cgroupv2
filesystem.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>