new(scap): cache v2 subsystem lookups
Since we now need to walk up the cgroup tree for every process, we
run a lot of redundant scans (the higher we go up the tree, the greater
the chance we've already parsed the cgroup.controllers file).

Introduce a simple cache, indexed by the full path to the cgroup
directory. The idea behind the somewhat weird clear/enable API is that
we only keep the cache enabled during repeated lookups (/proc scans)
and not for individual process lookups. The individual lookups may be
few and far apart and it's possible (however unlikely) that
the controllers will change across the hierarchy in the meantime.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>