feat(libsinsp): Track the last time a process called exec
The start time of a process is tracked by the last time it was cloned, however
this is a bit unintuitive in the case of exec as it would look like the new
executable started before the exec since the clone would have happened
beforehand.

Resolve this by adding a new threadinfo field m_lastexec_ts which tracks the
last time a process returned from an exec call. This way a user can either use
the last clone time or last exec time as appropriate for their use case. Note
that the exec time can be zero in the case where a process has not yet exec'ed.

The logic is implemented as follows:
* When we scan from /proc we initialise the last exec time to zero.
* On an exec exit event we update the last exec time for the process to the
  time of the event.
* On a thread clone we copy the last exec time from the parent, so all threads
  share the last exec time of the process.
* On a fork/process clone we set the last exec time to zero.

Signed-off-by: Allan Xavier <allan.xavier@sysdig.com>