When using pmatch, expand values via glob() first
Some users have asked for the ability to use glob expressions in
pmatch filterchecks, for example:

fd.name pmatch (/home/*/lib, /home/*/bin)

The idea here is that the directory below /home varies, but the
directories below each user directory does not.

To support this, expand any values passed to a CO_PMATCH filtercheck
via glob before adding them to the prefix_search struct. The memory
needs to be held by the filtercheck, so add a new vector of vectors
m_val_storages_paths_items to hold the expanded values.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>