Allow fd.* fields to work for fchmod/fchown by adding fdinfo to evt
fchmod/fchown act on a fd, but the fdinfo for the fd wasn't associated
with the event. That meant that fd.* fields wouldn't have values for
fchmod/fchown.

Fix this by adding a parser for fchmod/fchown exit events that grab
the fd, look it up from the threadinfo, and set m_fdinfo for the
event and tinfo->m_lastevent_fd for the associated threadinfo.

Also add EF_USES_FD to the fchmod/fchown exit events so the fd
filtercheck extract_fd will return true.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>