Make field output consistent wrt null termination
Depending on the specific syscall, the string representation of the
path might or might not end in a \0. Generally, ones populated by
evt.rawarg fields *will* end in a \0, as that value is provided
directly by the driver.

On the other hand, syscalls that use parsers to extract fields like
fchown/fchmod will *not* end in a \0.

As this field is all about consistency across syscalls, make the field
consisten by removing any final \0 character before returning the
extracted string.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>