fix(scap): initialize the platform before the engine
While it's cleaner to have the engine handle fully working before we
pass it to the platform init method, it's not strictly necessary:
the only use of the engine_handle is to store it in linux_platform
to call linux_vtable methods.

Initializing the platform before the engine fixes an issue with
the savefile engine, where we were putting data into the proclist
(in engine init) before the proclist callback was set (in platform init).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>