Remove ALWAYS_DROP setting for setsid system call
Libsinsp contains a parser for the SETSID_X event, and updates the m_sid
for the threadinfo object.  But the current ALWAYS_DROP setting for the
SETSID syscall_table entry, means this parsing never happens, in the
normal case where dropping mode is enabled in the kernel module/BPF probe.

This change removes the ALWAYS_DROP setting for SETSID, so applications
which rely on thread session IDs will have access to accurate information.

Signed-off-by: Joseph Pittman <joseph.pittman@sysdig.com>