fix(scap): fix platform initalization order (again)
While we need the proclist callback to be initialized before
the engine (which calls for initializing the platform first),
we also call linux_vtable methods (get_vtid) during the /proc scan
(which happens inside linux platform_init). That implies
initializing the engine first. Given this situation, we can only
split the platform setup into two parts:

- initializing the generic part (which engines may depend on)
- initializing the platform specific part (on which no engine should
  ever depend)

With the engine initialization in between the two.
656be921a5a6ab7cacb1cd4b370c1593a2c2a02f was a previous attempt to fix
the initialization order and this commit (mostly) reverts it.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>