fix(modern_bpf): fix NULL dereference in signal_deliver filler
The `signal_deliver` filler can be called with info=NULL
(`SEND_SIG_NOINFO`). Despite all I've been led to believe with eBPF,
this does cause an actual NULL dereference in the kernel,
promptly killing the machine (as the offending thread dies while
holding the spinlock in get_signal).

So let's check the pointer before we dereference it.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>
Co-Authored-By: Andrea Terzolo <andreaterzolo3@gmail.com>