fix: explicitly handle the `runc:[0:PARENT]` case
PPM_CL_CLONE_PARENT is set by `runc:[0:PARENT]` when it creates
the first process in the new pid namespace. In new sinsp versions
`PPM_CL_CHILD_IN_PIDNS` is enough but in old scap-files where we don't have
this custom flag, we leave the event to the child parser when `PPM_CL_CLONE_PARENT`
is set since we don't know if the new child is in a pid namespace or not.
Moreover when `PPM_CL_CLONE_PARENT` is set `PPM_CL_CLONE_NEWPID` cannot
be set according to the clone manual.

Signed-off-by: Andrea Terzolo <andreaterzolo3@gmail.com>