chore(scap): set up proc callbacks in platform alloc functions
The proc callbacks are the one part of the platform that needs to be
initialized very early (before the engine is open).

Move the proc callback setup to the allocation function so that we never
need to worry about the callback not getting set up early enough.

This isn't too pretty at an abstract level (we're mixing allocation
and a subset of initialization) but it's more practical this way.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>