build(minimal-build): allow using eBPF driver with minimal builds
Historically, the minimal build was only able to run using the kernel
module driver, but since the addition of the modern probe this has also
become an option for running with minimal builds. I believe there's
value in allowing adopters of the libs to use the minimal build with the
eBPF probe.

This change means that scap will always be linked against zlib and
libelf, but I believe this is a worthwhile trade off.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>