feature(scap): allow custom tracepoints on ebpf probe
This change allows eBPF probes created with tracepoints other than the
ones used by Falco to be directly attached and detached. This is useful
for adopters that might want to go through the additional effort of
attaching directly to the syscalls they care about, excluding sys_enter
and sys_exit which could add extra computing effort, even for ignored
syscalls. Because adopters need to go the extra mile to compile a probe
from their own source code, I don't think a separate mechanism for
controlling whether the custom tracepoints are attached or not is
needed, simply finding such a tracepoint means we want it attached.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>