build(modern_bpf): relax bpftool version test.
For most use cases, bpftool version checking works fine, However, when
using bpftool compiled from a kernel with backported patches, the
version might not be the most reliable way to see if the required
capabilities are avilable. As an example, RHEL 8.8 uses a bpftool with
versions that match 4.18, but because of it having backported patches
the modern probe can be compiled without errors. With this patch we move
from checking a specific version to trying to catch if bpftool supports
the gen command.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>