fix(libscap/engines): move some checks in `start_capture`
Doing these checks in the init method could cause some issues.
In the kernel module engine, an old driver with different IOCTL codes
could be loaded and if we use IOCTL codes before checking the
compatibility we could face a not user friendly error. For this reason,
we move this `mark_syscall` check when the `check_api_compatibility`
method is already called by scap.

Signed-off-by: Andrea Terzolo <andreaterzolo3@gmail.com>