fix(driver): use configure to fix build of RHEL8.9 kernels
RHEL8.9 kernels (4.18.0) have an inconsistent behavior in that
the first releases (up to ~500) do NOT include the backported
change, which was however included from a certain point on.
The previous implementation empirically (but wrongly) assumed
that all RHEL8.9 kernels would have the patch, so the first
ones would not compile.
Now that we have a dynamic discovery mechanism, we can drop
this assumption and leverage HAS_DEVNODE_ARG1_CONST.

Notice how this includes a revert of:

commit 977f946e31a2373501cc8a5ff8243632965a257b
Author: Gerlando Falauto <gerlando.falauto@sysdig.com>
Date:   Wed Aug 2 14:22:37 2023 +0200

    fix(driver): fix build on RHEL 8.9 kernels

Signed-off-by: Gerlando Falauto <gerlando.falauto@sysdig.com>
Co-authored-by: Angelo Puglisi <angelo.puglisi@sysdig.com>