fix(test/drivers): correct socketcall SYS_ACCEPT failures on s390x
The modern BPF driver maps the SYS_ACCEPT to SYS_ACCEPT4 which is being
tested in the socketcall SYS_ACCEPT tests.

The kmod and bpf drivers do not switch directly but rather ensure to
use the PPM event instead of syscall.  The socketcall SYS_ACCEPT tests
contain a specific handling for kmod.  Enable this handling to also for
the bpf driver.

Signed-off-by: Hendrik Brueckner <brueckner@de.ibm.com>