Add can_have_argument(), don't cache anything for fields w/ args
Add a can_have_argument() method that checks the flags for the field
for either EPF_ARG_REQUIRED/EPF_ARG_ALLOWED.

When can_have_argument() is true, skip caching of both extractions and
comparisons. This ensures that two fields that are the same but have
different arguments are not mistakenly cached together.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>