Use pid/tid stat file for accurate thread start time calculation
Current /proc scan logic uses the stat() system call of the cmdline file within
the /proc filesystem, to determine the start time of a thread or process.
This approach yields inaccurate and inconsistent results, because /proc files are
created on demand (e.g. /proc directory scan, /proc file open), rather than when
the process or thread is created.

The definitive approach is to read the system boot time from /proc/stat, then
add the start time reported by the /proc/<pid>/stat file for a process,
or the /proc/<pid>/task/<tid>/stat file for a non-process thread.

This changeset implements this definitive approach.

Signed-off-by: Joseph Pittman <joseph.pittman@sysdig.com>