fix(container/cri-engine): populate labels field for pod sandbox containers.
The CRI differentiates the normal containers from the pause container. The pause container
is the same as the "pod sandbox". In the CRI context the "pod sandbox" is the only place
where the labels of the pod (as in kubernetes pod) are set. When handling a "pod sandbox"
we populate the the labels field that are extracted when the "k8s.pod.labels" field is needed.

At the same time we make sure that the normal containers have a special label, "io.kubernetes.sandbox.id=podSandboxID".
The special labels is used by the filterchecks to retrive the right "pod sandbox" and then get its labels.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>