fix(driver): KERNELDIR always using current kernel
When installing a new kernel, dkms would build the module for it, but
with KERNELDIR defined with `uname -r` the compilation will likely fail.
Define it with the current kernel headers only when the makefile is used
directly, and rely on the make current directory otherwise (coming from
make -C ...).

Signed-off-by: Angelo Puglisi <angelopuglisi86@gmail.com>