fix(plugins): make list_fields return a const pointer
The fields returned by a plugin can be a static string and the plugin
framework doesn't have any business in modifying it so let's make it
const (to avoid a copy for plugins that can return a static string).

Note: I only bump the plugin API version by 0.0.1 since the changes
in this PR do not affect binary compatibility in any way.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>