fix(driver/modern): add always a null terminator after `args`
This commit handles `args` >= 4096. We always add a `\0` at the end of
the args even in case of partial reads. Moreover this changes uniform
the modern probe behavior with other drivers, before this commit we have
a max len of 4096 for `exe` and other `4096` for args. After this change
`exe + args` should be <= 4096

Signed-off-by: Andrea Terzolo <andreaterzolo3@gmail.com>