chore(docs): replace IN and OUT macros with docs
In scap.h, we have a couple of quite generic macros called IN and OUT
that can very easily leak into adopter's code bases and wreak havoc.

From my understanding, these macros are only used for documenting which
parameters in a function are input and which are output. Since they are
not consistently used, only appear in a handful of places in the
codebase and the preprocessor is the devil's spawn, we should replace
these macros with proper doxygen style string docs.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>
Co-authored-by: Olivier Valentin <ovalenti@redhat.com>