chore(metrics): refactor metrics v2 so it uses classes
This is an alternative to the original implementation. Instead of using
capturing lambdas that get called back after some additional work is
done, we use classes and gather data during instantiation of the object.
This approach should be a lot more straight forward and it also doesn't
create unneeded objects when there's no need to do so.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>