chore(ci): improve threshold checks.
Added a check on new leaks for heaptrack checks;
moreover, changed perf checks to run against
sum of diffs, that is more stable and useful.
Ie: if part of logic gets moved from one method to another,
we expect eg: method_A -> -7%, method_B -> +7%.
Before, this triggered the CI error; now it does not, since the sum is 0.

Signed-off-by: Federico Di Pierro <nierro92@gmail.com>