new(sinsp): add tools to manage plugin ABI compatibility
`extract_plugin_abi.py` parses the output from pahole describing
the `plugin_api` struct (with the offsets and sizes of all fields)
and generates a `plugin_abi.h` header with static asserts to ensure
the ABI is up to date (on x86_64/Linux).

`diff_plugin_abi.py` takes two copies of the generated plugin_abi.h
files (presumably, one from the repo and a freshly generated one)
and raises an error if:
 * a struct member has been moved/removed without ABI version bump
 * the ABI version has been increased without breaking changes

Note: adding a field at the end of the plugin_api struct does not
count as a breaking change.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>