fix(libscap): cgroup v2 path discovery fails with empty cgroup2 mountpoint
Fix the cgroup v2 path discovery mechanism in the presence of cgroup2
mount point with no controllers. The issue has been detected in a RKE
node, where the following empty mountpoint is breaking the detection
logic.

```
none on /host/var/run/calico/cgroup type cgroup2 (rw,relatime,nsdelegate,memory_recursiveprot)
```

The fix consists in stopping the iteration as soon as a mountpoint that
has available controllers listed in the `cgroup.controllers` is
detected.

Signed-off-by: Iacopo Rozzo <iacopo@sysdig.com>