fix(plugin): fix extract offsets handling
The docs (and a unit test) specified `value_offsets` to be an array
of `ss_plugin_extract_value_offsets` structs, while the code in
plugin_filtercheck.cpp expected it to be a struct of arrays.
Things worked out only because we never extract multiple fields
in one go (at least in libsinsp itself).

Make `value_offsets` an actual array of structs allocated by the caller
(not the plugin). Since the plugin needs to rely on the size of the
array now, make sure it's large enough (at least num_fields in size).

Additionally, clarify that the offsets are counted from the start
of the event buffer (including the header).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>