test(userspace/libsinsp/scap_files): add `assert_num_event_types()`
Replace helper to verify the presence of the exact number of events
of specific types inside a single scap file. Differently from the
old `assert_num_event_type()`, this helper is able to perform the
check on multiple event types at the same type: this helps
consolidating all the tests checking the number of events in all scap
file under a single test.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>