fix(userspace/libsinsp): mitigate memory leak related to clone*
The current implementation leverage two events to detect clone*
system call invocation and react to it. One event is associated with
the parent thread (caller), while the other with the child thread.
The implementation of the caller event parser is currently resilient
to possible child event drops, but doesn't correctly handle the
following possible events sequence:
- child event
- proc exit event
- caller event
This sequence is pretty common for short-lived threads.
In the current implementation, if this scenario happens, the caller
event parser don't find any child thread info, and (wrongly) create
it and add it to the thread table. This results in an unwanted thread
info being stuck in the thread table, and consuming memory until it
is removed by the threads auto-purging feature. In order to avoid the
temporary leak, this patch adds a circular buffer saving info about
the latest 1024 recently removed threads. Every time a thread is
removed, the parent tid and the tid are saved in the circular buffer:
this info is then leveraged in the caller event parser code to decide
if the thread should be added or not (because it was recently deleted)
to the thread table.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>
Co-authored-by: Leonardo Grasso <me@leonardograsso.com>
Co-authored-by: Iacopo Rozzo <iacopo@sysdig.com>