fix(userspace/libsinsp): avoid re-adding child thread in clone exit caller
The current implementation leverage two events to detect clone*
system call invocation and react to it. One event is associated with
the parent thread (caller), while the other with the child thread.
The implementation of the caller event parser is currently resilient
to possible child event drops, but doesn't correctly handle the
following possible events sequence:
- child event
- proc exit event
- caller event
This sequence is pretty common for short-lived threads.
In the current implementation, if this scenario happens, the caller
event parser don't find any child thread info, and (wrongly) create
it and add it to the thread table. This results in an unwanted thread
info being stuck in the thread table, and consuming memory until it
is removed by the threads auto-purging feature.

Furthemore, we know for sure that in some architectures (e.g., ARM),
the clone exit for the child was missing, and this logic here mitigated
this issue. However, it was later better addressed in the drivers,
using a different hook to catch the clone exit event.
For this reason, a better strategy is not to manage the child at this stage.
However, we still need to deal with old SCAP files that lack
child clone exit events, so we preserve the child addition logic
only when reading from SCAP files.

Signed-off-by: Leonardo Grasso <me@leonardograsso.com>
Co-authored-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>
Co-authored-by: Iacopo Rozzo <iacopo@sysdig.com>
Signed-off-by: Leonardo Grasso <me@leonardograsso.com>