feat(driver)!: add `EF_OLD_VERSION` to majority of enter events
As majority of enter events are not produced anymore by drivers, mark
them as old versions in the event table. There are still some events
which have not been marked yet, or will not be marked, for different
reasons. For reference, the list of non-marked enter event names is
the following:
- `PPME_SYSCALL_OPEN_E`
- `PPME_SOCKET_CONNECT_E`
- `PPME_DROP_E`
- `PPME_SCHEDSWITCH_6_E`
- `PPME_PROCEXIT_1_E`
- `PPME_SCAPEVENT_E`
- `PPME_SIGNALDELIVER_E`
- `PPME_PROCINFO_E`
- `PPME_CPU_HOTPLUG_E`
- `PPME_K8S_E`
- `PPME_TRACER_E`
- `PPME_MESOS_E`
- `PPME_NOTIFICATION_E`
- `PPME_INFRASTRUCTURE_EVENT_E`
- `PPME_PAGE_FAULT_E`
- `PPME_SYSCALL_OPENAT_2_E`
- `PPME_PLUGINEVENT_E`
- `PPME_CONTAINER_JSON_2_E`
- `PPME_SYSCALL_OPENAT2_E`
- `PPME_USER_ADDED_E`
- `PPME_USER_DELETED_E`
- `PPME_GROUP_ADDED_E`
- `PPME_GROUP_DELETED_E`
- `PPME_ASYNCEVENT_E`

BREAKING CHANGE: mark majority of enter events as old versions

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>