feat: add `C_ACTION_STORE_AND_SKIP` converter action support
Add `C_ACTION_STORE_AND_SKIP` converter action to indicate that a
copy of the event should be stored in the scap converter's internal
event storage and, at the same time, that another copy of the event
should proceed to the upper layers. As stated in the code, the current
implementation is identical to the `C_ACTION_STORE` one. However, in
the future, `C_ACTION_STORE` will return something to indicate that
the event should be dropped, while `C_ACTION_STORE_AND_SKIP` will
return `CONVERSION_SKIP` (or whatever is the name we choose to
indicate that the event should proceed... Maybe we can reuse
`CONVERSION_COMPLETE`). Update the implementation once we are ready to
change the `C_ACTION_STORE` logic.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>