refactor: use `C_ACTION_STORE_AND_SKIP` for {PPM_SOCKET_CONNECT_E, 2}
Use the scap converter action `C_ACTION_STORE_AND_SKIP` to state the
intention to store events with type `PPME_SOCKET_CONNECT_E` and 2
parameters in the scap converter's internal event storage and, at the
same time, send an event copy to the upper layers. This is a refactor
and not a feature because, currently, `C_ACTION_STORE_AND_SKIP`
implementation is equivalent to `C_ACTION_STORE`.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>