fix(userspace/libsinsp): fix extraction of the directory value
Extracting the value of a directory is broken and in some cases the
extraction returns also the filename. This will also impact Falco's
rule parsing because rules that depend on the correct value do no
longer work.

This fix makes sure that the code to strip off the leading filename
is run in all cases when the resulting string contains also
the filename.

Signed-off-by: Tero Kauppinen <tero.kauppinen@est.tech>