feat!: drop scap files' enter events not eligible for scap conversion
Drop scap files' enter events not eligible for scap conversion.
Moreover, formalize the conversion eligiblity rules both in code and
in a separate `README.md` file.

Conversion eligibility rules are used to determine if an event is
eligible for conversion, if it must be allowed to proceed towards
upper layers or must be dropped. Moreover, after each conversion,
these rules are always re-checked, to determine what to do with the
new event.

Static conversion eligibility rules are evaluated using a top-down
approach. They leverage the following definitions:

- an event having `EF_TMP_CONVERTER_MANAGED` among its flag is
  considered to be managed by the scap converter
- an event having `EF_OLD_VERSION` among its flag is considered an old
  event version

The following is the set of rules:

1. event not managed by the converter don't need any conversion;
   moreover
   1. new event versions are allowed to proceed towards upper layers
   2. old enter event versions are always dropped
   3. (validity check) exit events cannot be old without being managed
      by the converter
2. enter events always need a conversion
3. old exit event versions, or new exit event versions with a number
   of expected parameters (as specified in the event schema) different
   from their actual number of parameters, always need a conversion
4. all the other combinations don't need any conversion

BREAKING CHANGE: not eligible scap files' enter events are not
  delivered anymore to sinsp

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>