fix(ebpf): do not map overlapping perf event buffers
A fix for CVE-2025-38563 (https://github.com/torvalds/linux/commit/adf12a394c8e)
makes partially overlapping mappings of perf event buffers impossible.
Since we rely on precisely that, this breaks the eBPF driver on all kernels
with the fix.

This patch replaces the two overlapping mappings with just one plus some
scratch space after the ring buffer (64KiB). Whenever we're about to
consume an event that spans the end of the ring buffer, copy the missing
part from the beginning into the scratch space, so that the event is
contiguous in memory.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>