feat!: mark `{MESOS,TRACER,K8S}_E as old and `TRACER_X` as unused
Instead of changing the number of parameters for `TRACER_X` to 0, keep
the it set to 3, even if the event is now marked as unused. This is
not the perfect solution, but prevents breaking old scap files
readings: indeed, the savefile engine doesn't expect to read events,
encoded using a v1 event block, with less (inferred) parameters than
the ones declared in the event table.
In the future, we can adapt the engine code and remove any parameter
from `TRACER_X`. At the moment, the scap converter takes care of
droppping it regardless of its parameters number, since it is marked
as unused.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>