feat(userspace/libsinsp/parsers)!: drop redundant connect_x code
For both enter and exit events, `sinsp_parser::reset()` (which is
called before calling sinsp_parser::process_event()`), takes care of:
1) retrieving the connect's fd parameter
2) setting the last event fd for the event's thread info
3) retrieving the fd info associated to the fd
4) setting the event's fd info

This means that, if the connect exit event parsing code detects that
the event's fd info is null, there is nothing it can due to recover
the fd info, as an equal attempt has already been made by the
corresponding code in `sinsp_parser::reset()`. As a result, the code
trying to recover the event's fd info from the connect's fd parameter
is redundant and can be removed.

Since the connect exit event parser code cannot determine if the
returned fd info entry is stale or not, to be safe, always force it to
overwrite the connection's tuple destination

BREAKING CHANGE: destination is always overwritten

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>